update role and permission.

+++ commodity-service/src/main/java/com/xiaomi/commodity/entity/Category.java
@@ -15,12 +15,12 @@ public class Category implements Serializable {
 	/**
 	 * 类别Id
 	 */
-	private Long cateid;
+	private Integer cateid;
 
 	/**
 	 * 父类别id当id=0时说明是根节点,一级类别
 	 */
-	private Long parentid;
+	private Integer parentid;
 
 	/**
 	 * 类别名称
@@ -30,12 +30,12 @@ public class Category implements Serializable {
 	/**
 	 * 类别状态1-正常,2-已废弃
 	 */
-	private Long status;
+	private Integer status;
 
 	/**
 	 * 排序编号,同类展示顺序,数值相等则自然排序
 	 */
-	private Long sortorder;
+	private Integer sortorder;
 
 	/**
 	 * 创建时间
